Started the Spring 2006 PLClub schedule.
Fixed the unimplemented code for February in schedule.xslt, but just
ignoring leap years for now.  I should probably break out some of
the "date" handling code into a separate XSLT library.  It is kind of
unfortunate that the XSLT standard library is so impoverished.  Perhaps,
reimplement this in Xtatic, CDuce, Cw, or that new thing in C#?

Updated the makefile to reflect the new schedule.



git-svn-id: svn+ssh://ssh.seas.upenn.edu/project/cis/plclub1/svnroot/plclub-web@100 02941e39-d7f6-0310-8830-87495589a7ad